+++ CHANGELOG.md
@@ -1,1 +1,8 @@
-- Babel security fix
+# New Features
+
+- Ability to set custom headers (only in Node)
+- Ability to get custom headers
+- The serverside (^1.2) can now set cookies.
+  - New local events:
+    - onCookieSet
+    - onCookieDelete